+++ packages/conc/lib/Polysemy/Conc/Interpreter/Interrupt.hs
@@ -1,6 +1,6 @@
 {-# options_haddock prune #-}
 -- |Description: Interrupt interpreters
-module Polysemy.Conc.Interrupt where
+module Polysemy.Conc.Interpreter.Interrupt where
 
 import qualified Control.Concurrent.Async as A
 import Control.Concurrent.Async (AsyncCancelled)
@@ -14,10 +14,10 @@ import Polysemy.Internal.Tactics (liftT)
 import Polysemy.Time (Seconds (Seconds))
 import System.Posix.Signals (Handler (CatchInfoOnce, CatchOnce), SignalInfo, installHandler, keyboardSignal)
 
-import qualified Polysemy.Conc.Data.Critical as Critical
-import Polysemy.Conc.Data.Critical (Critical)
-import Polysemy.Conc.Data.Interrupt (Interrupt (..))
-import Polysemy.Conc.Data.Race (Race)
+import qualified Polysemy.Conc.Effect.Critical as Critical
+import Polysemy.Conc.Effect.Critical (Critical)
+import Polysemy.Conc.Effect.Interrupt (Interrupt (..))
+import Polysemy.Conc.Effect.Race (Race)
 import qualified Polysemy.Conc.Effect.Sync as Sync
 import Polysemy.Conc.Interpreter.Sync (interpretSync)
 import Polysemy.Conc.Race (race_)