+++ packages/conc/changelog.md
@@ -3,6 +3,7 @@
 * Change `Race.timeout` to take a `Sem` for the fallback instead of a pure value.
 * Export all `Queue` constructors from `Polysemy.Conc.Queue`.
 * Export all `Sync` constructors from `Polysemy.Conc.Sync`.
+* Move all interpreters to `Polysemy.Conc.Interpreter`.
 
 # 0.2.0.0
 * Add `read*` constructors for `Sync`